refactor: AssetsEditor component
