--- src/components/AssetEditor/index.ts
@@ -1,3 +1,0 @@
-import AssetEditor from './AssetEditor'
-
-export default AssetEditor
