--- src/components/AssetPackAssetsEditor/index.ts
@@ -1,3 +1,0 @@
-import AssetPackAssetsEditor from './AssetPackAssetsEditor'
-
-export default AssetPackAssetsEditor
