+++ src/components/AssetsEditor/AssetsEditor.types.ts
