+++ src/components/AssetsEditor/SingleAssetEditor/SingleAssetEditor.css
@@ -1,14 +1,14 @@
-.EditAsset {
+.SingleAssetEditor {
   display: flex;
 }
 
-.EditAsset .left-column {
+.SingleAssetEditor .left-column {
   display: flex;
   flex-direction: column;
   margin-right: 24px;
 }
 
-.EditAsset .metric {
+.SingleAssetEditor .metric {
   text-transform: uppercase;
   color: var(--secondary-text);
   font-size: 12px;
@@ -17,12 +17,12 @@
   align-items: center;
 }
 
-.EditAsset .metric .Icon {
+.SingleAssetEditor .metric .Icon {
   margin-right: 8px;
   margin-left: 4px;
 }
 
-.EditAsset .left-column img {
+.SingleAssetEditor .left-column img {
   width: 175px;
   height: 175px;
   border-radius: 10px;
@@ -30,29 +30,33 @@
   border: 2px solid #302e37;
 }
 
-.EditAsset .left-column img.ground {
+.SingleAssetEditor .left-column img.ground {
   padding: 0px;
 }
 
-.EditAsset .dcl.tag-field {
+.SingleAssetEditor .dcl.field {
+  margin-bottom: 10px;
+}
+
+.SingleAssetEditor .dcl.tag-field {
   height: 120px;
 }
 
-.EditAsset .dcl.tag-field .container {
+.SingleAssetEditor .dcl.tag-field .container {
   overflow-x: auto;
   width: 465px;
 }
 
-.EditAsset .dcl.tag-field .ui.dropdown {
+.SingleAssetEditor .dcl.tag-field .ui.dropdown {
   width: max-content;
   min-width: 100%;
 }
 
-.EditAsset .ui.radio.checkbox {
+.SingleAssetEditor .ui.radio.checkbox {
   margin-bottom: 20px;
 }
 
-.EditAsset .ui.radio.checkbox label {
+.SingleAssetEditor .ui.radio.checkbox label {
   color: var(--secondary-text);
   font-size: 13px;
   line-height: 18px;
@@ -61,16 +65,16 @@
   padding-left: 30px;
 }
 
-.EditAsset .ui.radio.checkbox label:before {
+.SingleAssetEditor .ui.radio.checkbox label:before {
   width: 16px;
   height: 16px;
 }
 
-.EditAsset .ui.radio.checkbox label:after {
+.SingleAssetEditor .ui.radio.checkbox label:after {
   width: 19px;
   height: 20px;
 }
 
-.EditAsset .dcl.select-field .ui.dropdown .text {
+.SingleAssetEditor .dcl.select-field .ui.dropdown .text {
   text-transform: capitalize;
 }
