+++ src/components/AssetsEditor/SingleAssetEditor/SingleAssetEditor.types.ts
@@ -1,5 +1,6 @@
 export type Props<T> = {
   asset: T
+  errors: Record<string, string>
   onChange(asset: T): void
 }
