+++ src/components/AssetsEditor/SingleAssetEditor/index.ts
@@ -1,0 +1,3 @@
+import SingleAssetEditor from './SingleAssetEditor'
+
+export default SingleAssetEditor
