+++ src/components/AssetsEditor/index.ts
@@ -1,0 +1,3 @@
+import AssetPackAssetsEditor from './AssetsEditor'
+
+export default AssetPackAssetsEditor
